+++ src/components/Switch/index.js
@@ -5,9 +5,10 @@ const Switch = ({
   onClick,
   elementId = 'switch1',
   checked,
+  testSection,
 }) => {
   return (
-    <div className='position--relative' data-oui-component={ true }>
+    <div className='position--relative' data-oui-component={ true } data-test-section={ testSection }>
       <input type="checkbox" id={ elementId } className="oui-switch" checked={ checked } onClick={ onClick }/>
       <label htmlFor={ elementId } data-on-label="On" data-off-label="Off"></label>
     </div>
@@ -21,6 +22,8 @@ Switch.propTypes = {
   elementId: PropTypes.string,
   // onClick function
   onClick: PropTypes.func,
+  /** Hook for automated JavaScript tests */
+  testSection: PropTypes.string,
 };
 
 export default Switch;